citra_qt: Fix per-game graphics api option
* This was a pain to implement due to the sheer amount of combinations. I hope there aren't any bugs...

+++ src/common/settings.h
@@ -447,7 +447,7 @@ struct Values {
 
     // Renderer
     SwitchableSetting<GraphicsAPI> graphics_api{GraphicsAPI::OpenGL, "graphics_api"};
-    SwitchableSetting<u16> physical_device{0, "physical_device"};
+    SwitchableSetting<u32> physical_device{0, "physical_device"};
     Setting<bool> renderer_debug{false, "renderer_debug"};
     Setting<bool> dump_command_buffers{false, "dump_command_buffers"};
     SwitchableSetting<bool> spirv_shader_gen{true, "spirv_shader_gen"};
